+++ client/front/src/components/Data/data.js
@@ -1,0 +1,48 @@
+const infoVuelos = [
+    {
+        "id": "TK16~30-TK198~1",
+        "departureTime": "23:50",
+        "arrivalTime": "19:45",
+        "duration": "31h 55m",
+        "airlinesNames": [
+            "Turkish Airlines"
+        ],
+        "stopoversCount": 1,
+        "price": "4789"
+    },
+    {
+        "id": "UA818~30-UA2675~31-UA7983~1",
+        "departureTime": "22:05",
+        "arrivalTime": "05:30",
+        "duration": "43h 25m",
+        "airlinesNames": [
+            "United Airlines"
+        ],
+        "stopoversCount": 2,
+        "price": "7293"
+    },
+    {
+        "id": "DL110~30-DL515~31-DL7~1",
+        "departureTime": "21:15",
+        "arrivalTime": "15:10",
+        "duration": "53h 55m",
+        "airlinesNames": [
+            "Delta Air Lines"
+        ],
+        "stopoversCount": 2,
+        "price": "7407"
+    },
+    {
+        "id": "LA2368~30-LA2478~31-LA7489~1",
+        "departureTime": "19:35",
+        "arrivalTime": "17:30",
+        "duration": "57h 55m",
+        "airlinesNames": [
+            "LATAM Airlines Group SA"
+        ],
+        "stopoversCount": 2,
+        "price": "5767"
+    }
+]
+
+export default infoVuelos
