--- client/front/src/components/Home.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import Card from './Card'
-
-function Home() {
-  return (
-    <div>
-        <Card />
-    </div>
-  )
-}
-
-export default Home
