+++ client/front/src/components/Home/Home.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Card from './Card/Card'
+
+function Home() {
+  return (
+    <div>
+        <Card />
+    </div>
+  )
+}
+
+export default Home
