+++ client/front/src/components/NavBar/NavBar.js
@@ -1,0 +1,5 @@
+
+
+export default function NavBar(){
+    
+}
